rework: frontend - category api service
for fetching category data for navbar was in previous form, now this data is fetching using api services
